Added .gitignore file
