+++ Currency_Exchange_Machine/App.xaml
@@ -1,0 +1,9 @@
+﻿<Application x:Class="Currency_Exchange_Machine.App"
+             xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
+             xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
+             xmlns:local="clr-namespace:Currency_Exchange_Machine"
+             StartupUri="MainWindow.xaml">
+    <Application.Resources>
+         
+    </Application.Resources>
+</Application>
