+++ Currency_Exchange_Machine/App.xaml.cs
@@ -1,0 +1,17 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Configuration;
+using System.Data;
+using System.Linq;
+using System.Threading.Tasks;
+using System.Windows;
+
+namespace Currency_Exchange_Machine
+{
+    /// <summary>
+    /// Logika interakcji dla klasy App.xaml
+    /// </summary>
+    public partial class App : Application
+    {
+    }
+}
