--- Currency_Exchange_Machine/Currency_Exchange_Machine.sln
@@ -1,25 +1,0 @@
-﻿
-Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio Version 17
-VisualStudioVersion = 17.7.34024.191
-MinimumVisualStudioVersion = 10.0.40219.1
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Currency_Exchange_Machine", "Currency_Exchange_Machine\Currency_Exchange_Machine.csproj", "{3997412F-1C92-44ED-80E5-EC137120ABA5}"
-EndProject
-Global
-	GlobalSection(SolutionConfigurationPlatforms) = preSolution
-		Debug|Any CPU = Debug|Any CPU
-		Release|Any CPU = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{3997412F-1C92-44ED-80E5-EC137120ABA5}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{3997412F-1C92-44ED-80E5-EC137120ABA5}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{3997412F-1C92-44ED-80E5-EC137120ABA5}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{3997412F-1C92-44ED-80E5-EC137120ABA5}.Release|Any CPU.Build.0 = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(SolutionProperties) = preSolution
-		HideSolutionNode = FALSE
-	EndGlobalSection
-	GlobalSection(ExtensibilityGlobals) = postSolution
-		SolutionGuid = {7E479B3D-7A8C-4ED6-8429-10B5DD662602}
-	EndGlobalSection
-EndGlobal
