--- Currency_Exchange_Machine/Currency_Exchange_Machine/MainWindow.xaml
@@ -1,53 +1,0 @@
-﻿<Window x:Class="Currency_Exchange_Machine.MainWindow"
-        xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
-        xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
-        xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
-        xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
-        xmlns:local="clr-namespace:Currency_Exchange_Machine"
-        mc:Ignorable="d"
-        Title="Currency Exchange Machine" Height="450" Width="800">
-    <Grid>
-        <Grid.ColumnDefinitions>
-            <ColumnDefinition Width="33*"/>
-            <ColumnDefinition Width="66*"/>
-            <ColumnDefinition Width="701*"/>
-        </Grid.ColumnDefinitions>
-        <Grid.Background>
-            <LinearGradientBrush EndPoint="0.5,1" StartPoint="0.5,0">
-                <GradientStop Color="Black" Offset="1"/>
-                <GradientStop Color="#FF1B2309"/>
-                <GradientStop Color="#FF1B2309" Offset="1"/>
-                <GradientStop Color="#FF1B2309" Offset="1"/>
-                <GradientStop Color="#FFA4E218" Offset="0.974"/>
-            </LinearGradientBrush>
-        </Grid.Background>
-        <Button Content="Calculate" HorizontalAlignment="Left" Margin="321,338,0,0" VerticalAlignment="Top" Click="Button_Click" FontSize="24" Height="54" Width="261" Background="#FF368E48" BorderThickness="0,1,1,1" SnapsToDevicePixels="True" RenderTransformOrigin="1,1" BorderBrush="#FF012B04" Foreground="#FF013802" FontFamily="Times New Roman" FontWeight="Bold" Grid.Column="2">
-            <Button.RenderTransform>
-                <TransformGroup>
-                    <ScaleTransform/>
-                    <SkewTransform/>
-                    <RotateTransform Angle="360.731"/>
-                    <TranslateTransform X="0"/>
-                </TransformGroup>
-            </Button.RenderTransform>
-        </Button>
-        <Label Content="Currency Exchange Machine &#xD;&#xA;    (USD TO PLN)" HorizontalAlignment="Left" Margin="-84,10,0,0" VerticalAlignment="Top" Background="{x:Null}" RenderTransformOrigin="0.5,0.5" Height="87" Width="992" HorizontalContentAlignment="Center" VerticalContentAlignment="Center" FontFamily="Times New Roman" FontSize="36" Foreground="#FFA5FF6A" FontStretch="Medium" Grid.ColumnSpan="3">
-            <Label.RenderTransform>
-                <TransformGroup>
-                    <ScaleTransform/>
-                    <SkewTransform/>
-                    <RotateTransform Angle="0.01"/>
-                    <TranslateTransform/>
-                </TransformGroup>
-            </Label.RenderTransform>
-        </Label>
-        <TextBox x:Name="textbox" HorizontalAlignment="Left" Margin="168,117,0,0" TextWrapping="Wrap" VerticalAlignment="Top" Width="266" TextChanged="TextBox_TextChanged" Height="76" FontSize="48" VerticalContentAlignment="Center" HorizontalContentAlignment="Center" Grid.Column="2"/>
-        <Label x:Name="label" Content=""  HorizontalAlignment="Left" Margin="223,266,0,0" VerticalAlignment="Top" FontFamily="Times New Roman" FontSize="20" Height="42" Width="156" HorizontalContentAlignment="Center" Grid.Column="2"/>
-        <Label Content="Dolary (USD):" HorizontalAlignment="Left" Margin="57,136,0,0" VerticalAlignment="Top" FontFamily="Times New Roman" FontSize="24" Background="{x:Null}" Width="157" Grid.Column="1" Grid.ColumnSpan="2"/>
-        <Label Content="WYNIK:" HorizontalAlignment="Left" Margin="253,234,0,0" VerticalAlignment="Top" FontFamily="Times New Roman" FontSize="22" Foreground="#FF033A09" FontWeight="Bold" Grid.Column="2"/>
-        <Label Content="lilykprogramming" HorizontalAlignment="Left" Margin="20,365,0,0" VerticalAlignment="Top" FontFamily="Times New Roman" FontSize="22" Height="39" Width="199" FontStyle="Italic" Grid.ColumnSpan="3"/>
-        <Label Content="US Dollars" HorizontalAlignment="Left" Margin="458,136,0,0" VerticalAlignment="Top" FontFamily="Times New Roman" FontSize="24" Grid.Column="2"/>
-        <Label Content="zł (Polish Zloty)" HorizontalAlignment="Left" Margin="399,268,0,0" VerticalAlignment="Top" FontFamily="Times New Roman" FontSize="24" Grid.Column="2"/>
-
-    </Grid>
-</Window>
